Separate image reading into its own file
